% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BUA}
\alias{BUA}
\title{BUA.}
\format{
A \code{sf} dataframe with only one \code{BUA} column for the corresponding \emph{ONS} codes.

You can easily add data loading for example the \url{https://github.com/lvalnegri/dmpkg.datauk} package.

For further details see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=all(BDY_BUA)}.
}
\usage{
BUA
}
\description{
Digital Vector Boundaries in \code{sf} format and \emph{WGS84} CRS (\emph{epsg}) 4326) 
for all 5,830 \emph{Built-up Areas} in the England And Wales.
}
\details{
Built by dissolving \code{OA} boundaries using the \code{output_areas} lookup table from the \url{https://github.com/lvalnegri/dmpkg.geouk} package.

Last update: 2015 (this is a \emph{frozen} entity).
}
\keyword{datasets}
