% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BUAS}
\alias{BUAS}
\title{BUAS}
\format{
A `sf` dataframe with only one `BUAS` column for the corresponding *ONS* codes.

You can easily add data loading for example the \url{https://github.com/lvalnegri/RdataUK} package.

For further details see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=all(BDY_BUAS)}.
}
\usage{
BUAS
}
\description{
Digital Vector Boundaries in `sf` format and *WGS84* CRS (*epsg* 4326) 
for all 1,826 *Built-up Areas Sub-divisions* in the England And Wales.
}
\details{
Built by dissolving `OA` boundaries using the `output_areas` lookup table from the \url{https://github.com/lvalnegri/RgeoUK} package.

Last update: 2015 (this is a *frozen* entity).
}
\keyword{datasets}
