% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CED}
\alias{CED}
\title{CED.}
\format{
A \code{sf} dataframe with only one \code{CED} column for the corresponding \emph{ONS} codes.

You can easily add data loading for example the \url{https://github.com/lvalnegri/dmpkg.datauk} package.

For further details see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=all(BDY_CED)}.
}
\usage{
CED
}
\description{
Digital Vector Boundaries in \code{sf} format and \emph{WGS84} CRS (\emph{epsg}) 4326) 
for all 1,717 \emph{County Electoral Divisions} in England.
}
\details{
Built by dissolving \code{OA} boundaries using the \code{output_areas} lookup table from the \url{https://github.com/lvalnegri/dmpkg.geouk} package.

Last update: MAY-21.
}
\keyword{datasets}
