% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CTY}
\alias{CTY}
\title{CTY.}
\format{
A \code{sf} dataframe with only one \code{CTY} column for the corresponding \emph{ONS} codes.

You can easily add data loading for example the \url{https://github.com/lvalnegri/dmpkg.datauk} package.

For further details see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=all(BDY_CTY)}.
}
\usage{
CTY
}
\description{
Digital Vector Boundaries in \code{sf} format and \emph{WGS84} CRS (\emph{epsg}) 4326) 
for all 90 \emph{Counties} (plus Local Authorities) in England (Scotland, N.Ireland, and Wales are present with a pseudo-code).
}
\details{
Built by dissolving \code{LAD} boundaries using the \code{output_areas} lookup table from the \url{https://github.com/lvalnegri/dmpkg.geouk} package.

Last update: MAY-21.
}
\keyword{datasets}
