% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmpkg.bnduk.R
\docType{package}
\name{dmpkg.bnduk}
\alias{dmpkg.bnduk}
\title{dmpkg.bnduk: This package contains the digital vector boundaries, in `sf` format and "simplified" at 20%, 
for main British geographies, (BGC) Generalised (20m), clipped to the coastline.}
