% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LSOA}
\alias{LSOA}
\title{LSOA.}
\format{
A SpatialPolygonsDataFrame with an \code{id} column

You can add data easily loading for example the package \url{https://github.com/lvalnegri/dmpkg.geodemouk}

More information about ??? boundaries can be found at \url{}

For further details on all the available UK Geographies (as of Nov 2020), see
\url{https://geoportal.statistics.gov.uk/datasets/register-of-geographic-codes-november-2020-for-the-united-kingdom}
}
\usage{
LSOA
}
\description{
Digital Vector Boundaries in \code{rds} format file for \em{Lower Layer Super Output Areas} in UK with geographical coordinates and
Built by dissolving \code{OA} boundaries using the \code{output_areas} lookup table from the \url{https://github.com/lvalnegri/dmpkg.geodemouk} package
}
\keyword{datasets}
