% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LSOA21}
\alias{LSOA21}
\title{LSOA21}
\format{
A `sf` dataframe with only one `LSOA21` column for the corresponding *ONS* codes.
}
\usage{
LSOA21
}
\description{
Digital Vector Boundaries in `sf` format and *WGS84* CRS (*epsg* 4326) 
for all 35,672 *Lower Layer Super Output Areas* in England and Wales.
}
\keyword{datasets}
