% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSOA21}
\alias{MSOA21}
\title{MSOA21}
\format{
A `sf` dataframe with only one `MSOA21` column for the corresponding *ONS* codes.
}
\usage{
MSOA21
}
\description{
Digital Vector Boundaries in `sf` format and *WGS84* CRS (*epsg* 4326) 
for all 7,264 *Medium Layer Super Output Areas* in England and Wales.
}
\keyword{datasets}
