% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OA}
\alias{OA}
\title{OA}
\format{
A `sf` dataframe with only one `OA` column for the corresponding *ONS* codes.

You can easily add data loading for example the \url{https://github.com/lvalnegri/RdataUK} package.

For further details see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=all(BDY_OA)}.
}
\usage{
OA
}
\description{
Digital Vector Boundaries in `sf` format and *WGS84* CRS (*epsg* 4326) 
for all 232,296 *Output Areas* in the UK (*Small Areas* in N.Ireland).
}
\details{
Last update: 2011 (this is a *frozen* entity).
}
\keyword{datasets}
