% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OA21}
\alias{OA21}
\title{OA21}
\format{
A `sf` dataframe with only one `OA21` column for the corresponding *ONS* codes.
}
\usage{
OA21
}
\description{
Digital Vector Boundaries in `sf` format and *WGS84* CRS (*epsg* 4326) 
for all 188,880 *Output Areas* in England and Wales.
}
\keyword{datasets}
