% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCA}
\alias{PCA}
\title{PCA}
\format{
A `sf` dataframe with only one `PCA` column for the corresponding *Royal Mail* codes.

You can easily add data loading for example the \url{https://github.com/lvalnegri/RdataUK} package.
}
\usage{
PCA
}
\description{
Digital Vector Boundaries in `sf` format and *WGS84* CRS (*epsg* 4326) 
for all 121 *Postcode Area* in the UK.
}
\details{
Built by dissolving `PCD` boundaries using the `output_areas` lookup table from the \url{https://github.com/lvalnegri/RgeoUK} package.

Last update: FEB-2022.
}
\keyword{datasets}
