% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCD}
\alias{PCD}
\title{PCD.}
\format{
A \code{sf} dataframe with only one \code{PCD} column for the corresponding \emph{Royal Mail} codes.

You can easily add data loading for example the \url{https://github.com/lvalnegri/dmpkg.datauk} package.
}
\usage{
PCD
}
\description{
Digital Vector Boundaries in \code{sf} format and \emph{WGS84} CRS (\emph{epsg}) 4326) 
for all 2,815 (active and geographical) \emph{Postcode Districts} in the UK.
}
\details{
Built by dissolving \code{PCS} boundaries using the \code{output_areas} lookup table from the \url{https://github.com/lvalnegri/dmpkg.geouk} package.

Last update: NOV-2021.
}
\keyword{datasets}
