% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCT}
\alias{PCT}
\title{PCT}
\format{
A `sf` dataframe with only one `PCT` column.

You can easily add data loading for example the \url{https://github.com/lvalnegri/RdataUK} package.
}
\usage{
PCT
}
\description{
Digital Vector Boundaries in `sf` format and *WGS84* CRS (*epsg* 4326) for all 1430 *Post Towns* in the UK.
}
\details{
Built by dissolving `PCD` boundaries using the `output_areas` lookup table from the \url{https://github.com/lvalnegri/RgeoUK} package.

Last update: FEB-2022.
}
\keyword{datasets}
