% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PFN}
\alias{PFN}
\title{PFN}
\format{
A `sf` dataframe with only one `PFN` column for the corresponding *ONS* codes.

You can easily add data loading for example the \url{https://github.com/lvalnegri/RdataUK} package.

For further details see \url{https://data.police.uk/data/boundaries/}.
}
\usage{
PFN
}
\description{
Digital Vector Boundaries in `sf` format and *WGS84* CRS (*epsg* 4326) 
for all TBD *Police Neighbourhoods* in the UK.
}
\details{
Built by dissolving `OA` boundaries using the `output_areas` lookup table from the \url{https://github.com/lvalnegri/RgeoUK} package.

Last update: DEC-21.
}
\keyword{datasets}
