% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RbndUK.R
\docType{package}
\name{RbndUK}
\alias{RbndUK}
\title{RbndUK: This package contains the digital vector boundaries, in `sf` format and "simplified" at 20%, 
for main British geographies, (BGC) Generalised (20m), clipped to the coastline.}
