% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RGN}
\alias{RGN}
\title{RGN}
\format{
A `sf` dataframe with only one `RGN` column for the corresponding *ONS* codes.

You can easily add data loading for example the \url{https://github.com/lvalnegri/RdataUK} package.

For further details see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=all(BDY_RGN)}.
}
\usage{
RGN
}
\description{
Digital Vector Boundaries in `sf` format and *WGS84* CRS (*epsg* 4326) 
for all 9 *Regions* in England (Scotland, N.Ireland, and Wales are present with a pseudo-code).
}
\details{
Built by dissolving `LAD` boundaries using the `output_areas` lookup table from the \url{https://github.com/lvalnegri/RgeoUK} package.

Last update: 2011 (this is a *frozen* entity).
}
\keyword{datasets}
