% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{STP}
\alias{STP}
\title{STP}
\format{
A `sf` dataframe with only one `STP` column for the corresponding *ONS* codes.

You can easily add data loading for example the \url{https://github.com/lvalnegri/RdataUK} package.

For further details see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=all(BDY_STP)}.
}
\usage{
STP
}
\description{
Digital Vector Boundaries in `sf` format and *WGS84* CRS (*epsg* 4326) 
for all 42 *Sustainability and Transformation Partnerships* in England.
}
\details{
Built by dissolving `CCG` boundaries using the `output_areas` lookup table from the \url{https://github.com/lvalnegri/RgeoUK} package.

Last update: .
}
\keyword{datasets}
