% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TTWA}
\alias{TTWA}
\title{TTWA}
\format{
A `sf` dataframe with only one `TTWA` column for the corresponding *ONS* codes.

You can easily add data loading for example the \url{https://github.com/lvalnegri/RdataUK} package.

For further details see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=all(BDY_TTWA)}.
}
\usage{
TTWA
}
\description{
Digital Vector Boundaries in `sf` format and *WGS84* CRS (*epsg* 4326) 
for all 228 *Travel To Work Areas* in the UK.
}
\details{
Built by dissolving `LSOA` boundaries using the `output_areas` lookup table from the \url{https://github.com/lvalnegri/RgeoUK} package.

Last update: 2011 (this is a *frozen* entity).
}
\keyword{datasets}
