% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WPZ}
\alias{WPZ}
\title{WPZ.}
\format{

}
\usage{
WPZ
}
\description{
Digital Vector Boundaries in \code{sf} format and \emph{WGS84} CRS (\emph{epsg}) 4326) for all 60,709 \emph{Workplace Zones} in the UK.
}
\details{
Last update: 2011 (this is a \emph{frozen} entity).
}
\keyword{datasets}
