% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CTY}
\alias{CTY}
\title{CTY}
\format{
A `sf` dataframe with only one `CTY` column for the corresponding *ONS* codes.

You can easily add data loading for example the \url{https://github.com/lvalnegri/RdataUK} package.

For further details see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=all(BDY_CTY)}.
}
\usage{
CTY
}
\description{
Digital Vector Boundaries in `sf` format and *WGS84* CRS (*epsg* 4326) 
for all 90 *Counties* (plus *Local Authorities*) in England (Scotland, N.Ireland, and Wales are present with a pseudo-code).
}
\details{
Built by dissolving `LAD` boundaries using the `output_areas` lookup table from the \url{https://github.com/lvalnegri/RgeoUK} package.

Last update: MAY-21.
}
\keyword{datasets}
