% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LSOA}
\alias{LSOA}
\title{LSOA.}
\format{
A \code{sf} dataframe with only one \code{LSOA} column for the corresponding \emph{ONS} codes.

You can easily add data loading for example the \url{https://github.com/lvalnegri/dmpkg.datauk} package.

For further details see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=all(BDY_OA)}.
}
\usage{
LSOA
}
\description{
Digital Vector Boundaries in \code{sf} format and \emph{WGS84} CRS (\emph{epsg}) 4326) 
for all 42,619 \emph{Lower Layer Super Output Areas} in the UK (\emph{Data Zones} in Scotland, \emph{Small Output Areas} in N.Ireland).
}
\details{
Built by dissolving \code{OA} boundaries using the \code{output_areas} lookup table from the \url{https://github.com/lvalnegri/dmpkg.geouk} package.

Last update: 2011 (this is a \emph{frozen} entity).
}
\keyword{datasets}
