% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LSOA}
\alias{LSOA}
\title{LSOA}
\format{
A `sf` dataframe with only one `LSOA` column for the corresponding *ONS* codes.

You can easily add data loading for example the \url{https://github.com/lvalnegri/RdataUK} package.

For further details see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=all(BDY_OA)}.
}
\usage{
LSOA
}
\description{
Digital Vector Boundaries in `sf` format and *WGS84* CRS (*epsg* 4326) 
for all 42,619 *Lower Layer Super Output Areas` in the UK (*Data Zones* in Scotland, *Small Output Areas* in N.Ireland).
}
\details{
Built by dissolving `OA` boundaries using the *output_areas* lookup table from the \url{https://github.com/lvalnegri/RgeoUK} package.

Last update: 2011 (this is a *frozen* entity).
}
\keyword{datasets}
