% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSOA}
\alias{MSOA}
\title{MSOA. Boundaries for Middle Layer Super Output Areas in UK with geographical coordinates and}
\format{
A SpatialPolygonsDataFrame with an \code{id} column

To add data load either the package \url{} or \url{}

More information about ??? boundaries can be found at \url{}

For further details on all the available UK Geographies (as of Nov 2020), see
\url{https://geoportal.statistics.gov.uk/datasets/register-of-geographic-codes-november-2020-for-the-united-kingdom}
}
\usage{
MSOA
}
\description{
MSOA. Boundaries for Middle Layer Super Output Areas in UK with geographical coordinates and
}
\keyword{datasets}
