% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OA}
\alias{OA}
\title{OA. Boundaries for Output Areas in UK with geographical coordinates and}
\format{
A SpatialPolygonsDataFrame with an \code{id} column

To add data load either the package \url{} or \url{}

For further details on the UK Geographies, see \url{http://}
}
\usage{
OA
}
\description{
OA. Boundaries for Output Areas in UK with geographical coordinates and
}
\keyword{datasets}
