% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OA}
\alias{OA}
\title{OA.}
\format{
A \code{sf} dataframe with only one \code{OA} column for the corresponding \em{ONS} codes

You can add data easily loading for example the \url{https://github.com/lvalnegri/dmpkg.datauk} package

For further details see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=all(BDY_OA)}
}
\usage{
OA
}
\description{
Digital Vector Boundaries in \code{sf} format file and \emph{WGS84} CRS (\emph{epsg}) 4326) 
for all 232,296 \em{Output Areas} in the UK (\em{Small Areas} in N.Ireland)
}
\keyword{datasets}
