% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PAR}
\alias{PAR}
\title{PAR.}
\format{
A \code{sf} dataframe with only one \code{PAR} column for the corresponding \emph{ONS} codes.

You can easily add data loading for example the \url{https://github.com/lvalnegri/dmpkg.datauk} package.

For further details see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=all(BDY_PAR)}.
}
\usage{
PAR
}
\description{
Digital Vector Boundaries in \code{sf} format and \emph{WGS84} CRS (\emph{epsg}) 4326) 
for all 11,561 \emph{Civil Parishes and Unparished Communities} in Great Britain.
}
\details{
Built by dissolving \code{OA} boundaries using the \code{output_areas} lookup table from the \url{https://github.com/lvalnegri/dmpkg.geouk} package.

Last update: .
}
\keyword{datasets}
