% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCT}
\alias{PCT}
\title{PCT.}
\format{
A \code{sf} dataframe with only one \code{PCT} column.

You can easily add data loading for example the \url{https://github.com/lvalnegri/dmpkg.datauk} package.
}
\usage{
PCT
}
\description{
Digital Vector Boundaries in \code{sf} format and \emph{WGS84} CRS (\emph{epsg}) 4326) 
for all 1430 \emph{Post Towns} in the UK.
}
\details{
Built by dissolving \code{PCD} boundaries using the \code{output_areas} lookup table from the \url{https://github.com/lvalnegri/dmpkg.geouk} package.

Last update: NOV-2021.
}
\keyword{datasets}
