% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WPZ}
\alias{WPZ}
\title{WPZ.}
\format{

}
\usage{
WPZ
}
\description{
Digital Vector Boundaries in \code{sf} format file and \emph{WGS84} CRS (\emph{epsg}) 4326) for all 60,709 \em{Workplace Zones} in the UK
}
\keyword{datasets}
