% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WPZ}
\alias{WPZ}
\title{WPZ}
\format{
A `sf` dataframe with only one `WPZ` column for the corresponding *ONS* codes.

You can easily add data easily loading for example the \url{https://github.com/lvalnegri/RdataUK} package.

For further details see the [ONS Open Geography Portal](https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=name&tags=all(BDY_WZ%2CDEC_2011)].
}
\usage{
WPZ
}
\description{
Digital Vector Boundaries in `sf` format and *WGS84* CRS (*epsg* 4326) for all 60,709 *Workplace Zones* in the UK.
}
\details{
Last update: 2011 (this is a *frozen* entity).
}
\keyword{datasets}
